+++ CHANGELOG.md
@@ -3,6 +3,44 @@
 All notable changes to this project will be documented in this file. This project adheres to
 [Semantic Versioning](http://semver.org/) and [this changelog format](http://keepachangelog.com/).
 
+## [1.0.0-alpha.4] - 2021-02-27
+
+### Added
+
+- The Eloquent schema now has `indexQuery` and `relatableQuery` methods. These allow filtering for authorization
+  purposes when a list of resources is being retrieved. For instance, it could filter those queries so that only models
+  belonging to the authenticated user are returned.
+- Can now determine whether multiple filters should return zero-to-one resource using the `Schema::isSingular()` method.
+
+### Changed
+
+- **BREAKING** The query builder classes have been updated for changes to the interfaces they implement. This adds
+  the `withRequest()` method and renames `using()` to `withQuery()`.
+- [#2](https://github.com/laravel-json-api/eloquent/issues/2) **BREAKING** The `Fillable` contract now type-hints the
+  request class in its read-only method signatures, and allows it to be `null`. The `ReadOnly` trait has been updated,
+  so this is unlikely to affect field classes if the trait has been used.
+- **BREAKING** If no request class is provided to the `ModelHydrator` class via its new  `withRequest()` method, it is
+  now assumed the hydration is occurring outside of a HTTP request. I.e. that the developer is manually triggering the
+  hydration. Without the HTTP request, fields will not be checked for their read-only state and will be filled if the
+  provided data has a value for the field. Implementing libraries must ensure that `withRequest()` is called when
+  filling values provided by a HTTP client.
+- **BREAKING** Renamed the `Builder` class to `JsonApiBuilder`. This change was made as it was confusing what
+  a `Builder` referred to, because Laravel uses this class name for Eloquent builders.
+- The `Relation::type()` method should now be used to set the inverse resource type on a relationship field.
+
+### Fixed
+
+- The `Builder::filters()` method now correctly yields both the schema's filter and the filters from a relationship, if
+  one is set. Previously the filters were not yielded correctly if there was a relationship.
+- the `QueryToOne` and `QueryToMany` builders now correctly use the model's relation name instead of the JSON:API field
+  name when retrieving the relationship object from the model. Previously this would have failed with an error if the
+  model relationship name was not the same as the JSON:API field name.
+
+### Deprecated
+
+- The `Relation::inverseType()` method is deprecated and will be removed in `1.0-stable`. Use `Relation::type()`
+  instead.
+
 ## [1.0.0-alpha.3] - 2021-02-09
 
 ### Added